backend/src/gps.py
from llama.agents import LlamaAgent
import mysql.connector

# Database configuration (similar to other scripts)
db_config = {
    'host': 'db',
    'user': 'root',
    'password': 'rootpassword',
    'database': 'mydatabase',
}

def optimize_paths():
    # Implement Llama or NLP integration here
    llama_agent = LlamaAgent()
    optimized_paths = llama_agent.optimize_paths()
    
    # Store optimized paths in database (example implementation)
    store_optimized_paths(optimized_paths)

    return optimized_paths

def store_optimized_paths(optimized_paths):
    # Example implementation to store data in database
    db_connection = mysql.connector.connect(**db_config)
    cursor = db_connection.cursor()

    # Assume optimized_paths is structured data ready to be stored in your database tables
    # Example: Insert optimized paths into a new table or update existing tables

    cursor.close()
    db_connection.close()

def run_gps_agent():
    optimized_paths = optimize_paths()
    return optimized_paths
