backend/requirements.txt
llama-agents
llama-index-agent-openai
flask 
llama-cpp-python 
pandas
mysql-connector-python